Add CUDA to the image building, and include it in the install script

+++ grid-image-build/create-centos-image.sh
@@ -3,6 +3,11 @@ orgimage=${1}
 image='centos-grid.qcow2'
 rawimage='centos-grid.raw'
 
+if [ ! -f cuda.run ]; then
+  echo "[ERROR] du mangler cuda.run i denne mappa"
+  exit 1
+fi
+
 echo "Copying image"
 cp $orgimage $image
 echo "Upgrading packages"
@@ -12,7 +17,11 @@ virt-customize -a $image --install epel-release
 echo "Installing packages"
 virt-customize -a $image --install gcc,dkms,make,kernel-devel,cpp,glibc-devel,glibc-headers,kernel-headers,libmpc,mpfr
 echo "Installing GRID-script"
-virt-customize -a $image --copy-in check-grid-driver.sh:/opt/ --append-line '/etc/crontab:@reboot root bash /opt/check-grid-driver.sh'
+virt-customize -a $image \
+  --copy-in check-grid-driver.sh:/opt/ \
+  --copy-in cuda.run:/opt/ \
+  --copy-in cuda.sh:/etc/profile.d/ \
+  --append-line '/etc/crontab:@reboot root bash /opt/check-grid-driver.sh'
 echo "Re-labling selinux"
 virt-customize -a $image --selinux-relabel
