+++ grid-image-build/create-ubuntu-image.sh
@@ -3,6 +3,11 @@ orgimage=${1}
 image='ubuntu-grid.qcow2'
 rawimage='ubuntu-grid.raw'
 
+if [ ! -f cuda.run ]; then
+  echo "[ERROR] du mangler cuda.run i denne mappa"
+  exit 1
+fi
+
 echo "Copying image"
 cp $orgimage $image
 echo "Upgrading packages"
@@ -10,7 +15,11 @@ virt-customize -a $image --update
 echo "Installing packages"
 virt-customize -a $image --install build-essential,dkms,libxml2-utils
 echo "Installing GRID-script"
-virt-customize -a $image --copy-in check-grid-driver.sh:/opt/ --append-line '/etc/crontab:@reboot root bash /opt/check-grid-driver.sh'
+virt-customize -a $image \
+  --copy-in check-grid-driver.sh:/opt/ \
+  --copy-in cuda.run:/opt/ \
+  --copy-in cuda.sh:/etc/profile.d/ \
+  --append-line '/etc/crontab:@reboot root bash /opt/check-grid-driver.sh'
 
 echo "Converting to raw"
 qemu-img convert -f qcow2 -O raw $image $rawimage
